+++ src/MoBi.Presentation/Presenter/Main/MenuAndToolBarPresenter.cs
@@ -222,9 +222,6 @@ namespace MoBi.Presentation.Presenter.Main
             _menuBarItemRepository[MenuBarItemIds.NewIndividual].Enabled = enabled;
             _menuBarItemRepository[MenuBarItemIds.NewExpressionProfile].Enabled = enabled;
             _menuBarItemRepository[MenuBarItemIds.NewModule].Enabled = enabled;
-            _menuBarItemRepository[MenuBarItemIds.NewMetabolizingEnzyme].Enabled = enabled;
-            _menuBarItemRepository[MenuBarItemIds.NewTransportProtein].Enabled = enabled;
-            _menuBarItemRepository[MenuBarItemIds.NewSpecificBindingPartner].Enabled = enabled;
             _menuBarItemRepository[MenuBarItemIds.EditProjectSimulationSettings].Enabled = enabled;
             _menuBarItemRepository[MenuBarItemIds.SaveProjectSimulationSettings].Enabled = enabled;
             _menuBarItemRepository[MenuBarItemIds.LoadProjectSimulationSettings].Enabled = enabled;
